update(permissions): :zap: update guard and decorator for using with role base
Read `CHANGELOG.md` for more details
