+++ apps/api/src/controllers/index.ts
@@ -1,7 +1,7 @@
 import { AppController } from './app.controller';
 import { ProductsController } from './products.controller';
 import { AuthController } from './auth.controller';
-import { OrderController } from './order.controller';
+import { OrdersController } from './orders.controller';
 import { UsersController } from './users.controller';
 import { AttributesController } from './attributes.controller';
 import { CategoriesController } from './categories.controller';
@@ -28,7 +28,7 @@ export const ListControllers = [
     AttributesController,
     CartsController,
     AddressController,
-    OrderController,
+    OrdersController,
     OrdersManagementController,
     UsersController,
     // DiscountsController,
