+++ apps/api/src/controllers/stats.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Inject, Query, UseGuards, Headers } from '@nestjs/common';
+import { Controller, Get, Inject, Query, Headers } from '@nestjs/common';
 import { ClientRMQ } from '@nestjs/microservices';
 import { ACCESS_TOKEN_NAME, MANAGEMENTS_SERVICE } from '~libs/common/constants';
 import {
@@ -14,12 +14,13 @@ import {
     ApiTooManyRequestsResponse,
     ApiUnauthorizedResponse,
 } from '@nestjs/swagger';
-import { StaffGuard } from '~libs/common';
 import { StatsMntMessagePattern } from '~apps/managements/stats-mnt/stats-mnt.pattern';
 import { GetStatsRequestDTO, GetStatsResponseDTO } from '~apps/managements/stats-mnt/dtos';
 import { sendMessagePipeException } from '~libs/common/RabbitMQ/rmq.util';
 import { THeaders } from '~libs/common/types/common.type';
 import { GetStatsOrdersApiRequestDTO } from '~apps/managements/stats-mnt/dtos/get-stats-orders-request.2.dto';
+import { UserRole } from '~libs/resource/users/enums';
+import { Auth } from '~libs/common/decorators';
 
 @ApiBadRequestResponse({
     description: 'Invalid request, please check your request data!',
@@ -40,7 +41,7 @@ import { GetStatsOrdersApiRequestDTO } from '~apps/managements/stats-mnt/dtos/ge
     description: 'Internal server error, please try again later!',
 })
 @ApiBearerAuth(ACCESS_TOKEN_NAME)
-@UseGuards(StaffGuard)
+@Auth(UserRole.Manager, UserRole.Staff)
 @ApiTags('stats')
 @Controller('stats')
 export class StatsController {
