--- libs/common/src/constants/auth.constant.ts
@@ -1,4 +1,0 @@
-export const SKIP_AUTH_SUPER_ADMIN_GUARD = 'SKIP_AUTH_SUPER_ADMIN_GUARD';
-export const SKIP_AUTH_ADMIN_GUARD = 'SKIP_AUTH_ADMIN_GUARD';
-export const SKIP_AUTH_MOD_GUARD = 'SKIP_AUTH_MOD_GUARD';
-export const SKIP_AUTH_GUARD = 'SKIP_AUTH_GUARD';
