+++ libs/common/src/decorators/index.ts
@@ -1,3 +1,5 @@
 export * from './current-user.decorator';
 export * from './multi-file.decorator';
 export * from './file.decorator';
+export * from './required-role.decorator';
+export * from './auth.decorator';
