+++ libs/common/src/decorators/required-role.decorator.ts
@@ -1,0 +1,4 @@
+import { SetMetadata } from '@nestjs/common';
+import { UserRole } from '~libs/resource/users/enums';
+
+export const RequiredRoles = (...roles: UserRole[]) => SetMetadata(RequiredRoles.name, roles);
