+++ libs/common/src/guards/index.ts
@@ -1,5 +1,2 @@
 export * from './auth.guard';
-export * from './staff.guard';
-export * from './manager.guard';
-export * from './user.guard';
 export * from './optional-auth.guard';
