+++ libs/common/src/types/current-user.type.ts
@@ -1,6 +1,8 @@
 import { Types } from 'mongoose';
+import { UserRole } from '~libs/resource/users/enums';
 
 export type TCurrentUser = {
     _id: Types.ObjectId | string;
+    role: UserRole | string;
     [key: string]: any;
 };
